Rename the layer folder to layers to be consistent with other similar packages (#576)
