+++ doc/examples/layers/lib/main.dart
@@ -2,7 +2,7 @@ import 'package:flame/extensions/vector2.dart';
 import 'package:flame/game.dart';
 import 'package:flutter/material.dart' hide Animation;
 import 'package:flame/sprite.dart';
-import 'package:flame/layer/layer.dart';
+import 'package:flame/layers/layer.dart';
 import 'package:flame/flame.dart';
 
 import 'dart:ui';
